only 3-4 filters/files left to be done working through filters to move field class to map and make consistent naming scheme

+++ Source/DREAM3DLib/GenericFilters/FindNeighbors.h
@@ -73,11 +73,11 @@ class DREAM3DLib_EXPORT FindNeighbors : public AbstractFilter
 
   private:
     int32_t* m_GrainIds;
-    int32_t* m_SurfaceVoxels;
+    int8_t* m_SurfaceVoxels;
     bool* m_SurfaceFields;
     int32_t* m_Phases;
     int32_t* m_NumNeighbors;
-    int32_t* m_TotalSurfaceArea;
+    float* m_TotalSurfaceArea;
     NeighborList<int>* m_NeighborList;
     NeighborList<float>* m_SharedSurfaceAreaList;
